requirements.txt
pickle
pandas
joblib
google-generativeai
xgboost